# src/swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0",
    "title": "platzi",
    "contact": {}
  },
  "host": "localhost:3000",
  "basePath": "/api",
  "schemes": [
    "https"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/categories": {
      "get": {
        "summary": "get categories",
        "tags": [
          "Misc"
        ],
        "operationId": "getcategories",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        }
      },
      "post": {
        "summary": "post categories",
        "tags": [
          "Misc"
        ],
        "operationId": "postcategories",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "consumes": [
          "application/x-www-form-urlencoded"
        ],
        "parameters": [
          {
            "name": "name",
            "in": "formData",
            "required": true,
            "type": "string",
            "description": ""
          },
          {
            "name": "img",
            "in": "formData",
            "required": true,
            "type": "string",
            "description": ""
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        }
      }
    },
...
  },
...
}